requirements.txt
# Use python 3.8
drake
numpy
matplotlib
pygame